refactor: split results.rs into a results/ output module
Closes #1855. Port of the pro edition's split (hayabusa-pro#176). `src/results.rs`
(~4,092 lines, ~8 concerns) is split into a `src/results/` module directory
(mod.rs style, matching src/detections/) with focused submodules:

- src/results/mod.rs        - writer construction, output_results(_inner),
                              sort/dedup, ResultOutputState, the test module
- src/results/csv.rs        - emit_csv / emit_csv_inner
- src/results/json.rs       - json_scalar/json_string, JsonNode, group_details_raw,
                              serialize_record_body, output_json_str, ...
- src/results/display.rs    - _get_serialized_disp_output, _format_cellpos, ColPos
- src/results/summary.rs    - calc_statistic_info, output_result_summary, the
                              _print_* summary renderers, rule-author aggregation
- src/results/html_stock.rs - _output_html_computer_by_mitre_attck

Pure code movement: no function body, literal, or logic was changed. The public
API (emit_csv, output_json_str, output_result_summary, ...) is re-exported from
mod.rs so `crate::results::NAME` keeps resolving; the test module is unchanged.
(The submodule is `csv` so `use ::csv::...` disambiguates the extern crate.)

Verified byte-identical output on the full sample-evtx corpus (599 evtx, all
rules, sorted): csv-timeline, json-timeline (-o) and json-timeline (-L) are all
byte-for-byte identical to main. Full lib test suite (478) passes; cargo fmt
--check and cargo clippy --all-targets -- -D warnings clean.
