+++ src/results/mod.rs
@@ -1,35 +1,18 @@
-use std::cmp::{self, Ordering, min};
+use std::cmp::Ordering;
 use std::error::Error;
 use std::fs::File;
-use std::io::{self, BufWriter, Write};
+use std::io::{self, BufWriter};
 use std::process;
-use std::str::FromStr;
 
-use aho_corasick::{AhoCorasick, AhoCorasickBuilder, MatchKind};
+use ::csv::{QuoteStyle, Writer, WriterBuilder};
 use chrono::{DateTime, Local, TimeZone, Utc};
-use comfy_table::modifiers::UTF8_ROUND_CORNERS;
-use comfy_table::presets::UTF8_FULL;
-use comfy_table::*;
 use compact_str::CompactString;
-use csv::{QuoteStyle, Writer, WriterBuilder};
 use hashbrown::{HashMap, HashSet};
-use indexmap::IndexMap;
-use itertools::Itertools;
-use krapslog::{build_sparkline, build_time_markers};
-use nested::Nested;
-use num_format::{Locale, ToFormattedString};
-use serde::ser::{Serialize, SerializeMap, Serializer};
 use strum::IntoEnumIterator;
-use termcolor::{Buffer, BufferWriter, Color, ColorChoice, ColorSpec, WriteColor};
-use terminal_size::Width;
-use terminal_size::terminal_size;
+use termcolor::{Buffer, BufferWriter, ColorChoice, ColorSpec, WriteColor};
 
-use crate::detections::configs::{
-    Action, CURRENT_EXE_PATH, GEOIP_DB_PARSER, StoredStatic, TimeFormatOptions,
-};
-use crate::detections::message::{
-    AlertMessage, COMPUTER_MITRE_ATTCK_MAP, COMPUTER_MITRE_ATTCK_UNIQUE_KEYS, DetectInfo,
-};
+use crate::detections::configs::{Action, StoredStatic, TimeFormatOptions};
+use crate::detections::message::{AlertMessage, DetectInfo};
 
 /// Escapes user-supplied data values before embedding them into the Markdown that
 /// is rendered as HTML for the report, to prevent XSS and Markdown injection.
@@ -45,7 +28,7 @@ use crate::detections::message::{
 /// This should be applied only to dynamic data values (e.g. computer names from
 /// logs), not to the Markdown template strings that may contain intentional HTML
 /// like `<br>` or intentional Markdown like table pipes.
-fn html_escape_value(s: &str) -> String {
+pub(crate) fn html_escape_value(s: &str) -> String {
     let mut escaped = String::with_capacity(s.len());
     for c in s.chars() {
         match c {
@@ -69,13 +52,28 @@ fn html_escape_value(s: &str) -> String {
     }
     escaped
 }
-use crate::detections::utils::{
-    self, check_setting_path, format_time, get_writable_color, output_and_data_stack_for_html,
-    write_color_buffer,
-};
-use crate::level::{_get_output_color, LEVEL, create_output_color_map};
-use crate::options::htmlreport;
+use crate::level::LEVEL;
 use crate::options::profile::Profile;
+
+mod csv;
+mod display;
+mod html_stock;
+mod json;
+mod summary;
+
+use csv::emit_csv_inner;
+use summary::calc_statistic_info;
+
+pub use csv::emit_csv;
+pub use json::output_json_str;
+pub use summary::output_result_summary;
+
+#[cfg(test)]
+pub(crate) use crate::detections::utils::format_time;
+#[cfg(test)]
+pub(crate) use json::{close_unterminated_quote, json_scalar, json_string};
+#[cfg(test)]
+pub(crate) use summary::_print_unique_results;
 
 /// The same display color expressed once for the termcolor crate (plain terminal output) and
 /// once for the comfy_table crate (table output).
@@ -181,7 +179,7 @@ impl Default for ResultOutputState {
 /// the target (records separated by a newline — `first` tracks whether one has been written yet —
 /// reproducing the byte layout the CSV writer's `\n` delimiter previously produced, without
 /// abusing it as a plain-text sink).
-enum ResultWriter {
+pub(crate) enum ResultWriter {
     // Boxed: a `csv::Writer` is much larger than the `Json` variant.
     Csv(Box<Writer<Box<dyn io::Write>>>),
     Json {
@@ -195,9 +193,9 @@ enum ResultWriter {
 /// directly). `display_flag` is true when no output file was specified, i.e. results are
 /// displayed on the terminal.
 pub struct OutputWriter {
-    display_writer: BufferWriter,
-    disp_wtr_buf: Buffer,
-    result_writer: ResultWriter,
+    pub(crate) display_writer: BufferWriter,
+    pub(crate) disp_wtr_buf: Buffer,
+    pub(crate) result_writer: ResultWriter,
     pub display_flag: bool,
 }
 
@@ -273,36 +271,7 @@ pub fn output_results(
     }
 }
 
-/// Writes one batch of detections and folds it into the summary statistics. This is the
-/// streaming output path used in low-memory mode, where results are emitted per batch instead
-/// of being collected, sorted, and written all at once by `output_results`. Exits the process
-/// if writing fails.
-pub fn emit_csv(
-    detect_infos: &[DetectInfo],
-    duplicate_indices: &HashSet<usize>,
-    stored_static: &StoredStatic,
-    output_writer: &mut OutputWriter,
-    result_state: &mut ResultOutputState,
-) {
-    if detect_infos.is_empty() {
-        return;
-    }
-
-    let result = emit_csv_inner(
-        detect_infos,
-        duplicate_indices,
-        stored_static,
-        output_writer,
-        result_state,
-    );
-    if result.is_err() {
-        handle_output_error(Box::new(result.err().unwrap()));
-    }
-
-    calc_statistic_info(detect_infos, duplicate_indices, result_state, stored_static);
-}
-
-fn handle_output_error(err: Box<dyn Error>) {
+pub(crate) fn handle_output_error(err: Box<dyn Error>) {
     AlertMessage::alert(&format!("Failed to write CSV. {err}")).ok();
     process::exit(1);
 }
@@ -351,749 +320,6 @@ fn output_results_inner(
     output_result_summary(stored_static, output_writer, result_state);
 
     Ok(())
-}
-
-fn emit_csv_inner(
-    detect_infos: &[DetectInfo],
-    duplicate_indices: &HashSet<usize>,
-    stored_static: &StoredStatic,
-    output_writer: &mut OutputWriter,
-    result_state: &mut ResultOutputState,
-) -> io::Result<()> {
-    // Field values reach here with real `\n`/`\r`/`\t` (kept by utils::remove_sp_char). For
-    // single-line output the whole value is whitespace-collapsed and output_remover flattens any
-    // remaining control characters to spaces; with the multiline (or tab-separator) option the
-    // "🛂🛂" marker used to join multi-valued entries and the " ¦ " field separator become line
-    // breaks (or tabs) instead. Keep the (pattern, replacement) pairs in a single ordered list so
-    // the automaton's patterns and the replacement slice line up — `AhoCorasick::replace_all`
-    // indexes the replacements by pattern id.
-    let mut removed_replaced_pairs: Vec<(&str, &str)> = vec![("\n", " "), ("\r", " "), ("\t", " ")];
-    if stored_static.multiline_flag {
-        removed_replaced_pairs.push(("🛂🛂", "\r\n"));
-        removed_replaced_pairs.push((" ¦ ", "\r\n"));
-    } else if stored_static.tab_separator_flag {
-        removed_replaced_pairs.push(("🛂🛂", "\t"));
-        removed_replaced_pairs.push((" ¦ ", "\t"));
-    }
-    let output_remover = AhoCorasickBuilder::new()
-        .match_kind(MatchKind::LeftmostLongest)
-        .build(removed_replaced_pairs.iter().map(|(pattern, _)| *pattern))
-        .unwrap();
-    let remover_vals: Vec<&str> = removed_replaced_pairs
-        .iter()
-        .map(|(_, replacement)| *replacement)
-        .collect();
-
-    // Prepare the per-level color map, then determine the output format and whether duplicate
-    // field data should be replaced with "DUP" (the remove-duplicate-data option).
-    let color_map = create_output_color_map(stored_static.common_options.no_color);
-    let (json_output_flag, jsonl_output_flag, remove_duplicate_data) =
-        match &stored_static.config.action.as_ref().unwrap() {
-            Action::JsonTimeline(option) => (
-                true,
-                option.jsonl_timeline,
-                option.output_options.remove_duplicate_data,
-            ),
-            Action::CsvTimeline(option) => {
-                (false, false, option.output_options.remove_duplicate_data)
-            }
-            _ => (false, false, false),
-        };
-
-    let profile = stored_static.profiles.as_ref().unwrap();
-    for (i, detect_info) in detect_infos.iter().enumerate() {
-        if duplicate_indices.contains(&i) {
-            continue;
-        }
-        if output_writer.display_flag && !(json_output_flag || jsonl_output_flag) {
-            // Terminal display output.
-            if !result_state.has_displayed_header {
-                // Print the header row only once.
-                _get_serialized_disp_output(
-                    &output_writer.display_writer,
-                    profile,
-                    true,
-                    (&output_remover, &remover_vals),
-                    stored_static.common_options.no_color,
-                    get_writable_color(
-                        _get_output_color(&color_map, &detect_info.level),
-                        stored_static.common_options.no_color,
-                    ),
-                );
-                result_state.has_displayed_header = true;
-            }
-            _get_serialized_disp_output(
-                &output_writer.display_writer,
-                &detect_info.output_fields,
-                false,
-                (&output_remover, &remover_vals),
-                stored_static.common_options.no_color,
-                get_writable_color(
-                    _get_output_color(&color_map, &detect_info.level),
-                    stored_static.common_options.no_color,
-                ),
-            );
-        } else if jsonl_output_flag {
-            // JSONL output format
-            let result = output_json_str(
-                detect_info,
-                result_state,
-                jsonl_output_flag,
-                GEOIP_DB_PARSER.read().unwrap().is_some(),
-                remove_duplicate_data,
-            );
-            result_state.prev_message = result.1;
-            result_state
-                .prev_details_convert_map
-                .clone_from(&detect_info.details_convert_map);
-            if output_writer.display_flag {
-                write_color_buffer(
-                    &output_writer.display_writer,
-                    None,
-                    &format!("{{ {} }}", &result.0),
-                    true,
-                )
-                .ok();
-            } else if let ResultWriter::Json { writer, first } = &mut output_writer.result_writer {
-                // Separate records with a newline (the previous csv-writer delimiter), without a
-                // leading or trailing one.
-                if !*first {
-                    writer.write_all(b"\n")?;
-                }
-                *first = false;
-                write!(writer, "{{ {} }}", &result.0)?;
-            }
-        } else if json_output_flag {
-            // JSON output
-            let result = output_json_str(
-                detect_info,
-                result_state,
-                jsonl_output_flag,
-                GEOIP_DB_PARSER.read().unwrap().is_some(),
-                remove_duplicate_data,
-            );
-            result_state.prev_message = result.1;
-            result_state
-                .prev_details_convert_map
-                .clone_from(&detect_info.details_convert_map);
-            if output_writer.display_flag {
-                write_color_buffer(
-                    &output_writer.display_writer,
-                    None,
-                    &format!("{{\n{}\n}}", &result.0),
-                    true,
-                )
-                .ok();
-            } else if let ResultWriter::Json { writer, first } = &mut output_writer.result_writer {
-                // The previous csv writer joined the "{", body and "}" fields with its `\n`
-                // delimiter; reproduce that layout, with records separated by a newline.
-                if !*first {
-                    writer.write_all(b"\n")?;
-                }
-                *first = false;
-                write!(writer, "{{\n{}\n}}", &result.0)?;
-            }
-        } else if let ResultWriter::Csv(csv_writer) = &mut output_writer.result_writer {
-            // CSV output format
-            if !result_state.has_displayed_header {
-                csv_writer.write_record(detect_info.output_fields.iter().map(|x| x.0.trim()))?;
-                result_state.has_displayed_header = true;
-            }
-            csv_writer.write_record(detect_info.output_fields.iter().map(|x| {
-                match x.1 {
-                    Profile::Details(_) | Profile::AllFieldInfo(_) | Profile::ExtraFieldInfo(_) => {
-                        let ret = if remove_duplicate_data
-                            && x.1.to_value()
-                                == result_state
-                                    .prev_message
-                                    .get(&x.0)
-                                    .unwrap_or(&Profile::Literal("-".into()))
-                                    .to_value()
-                        {
-                            "DUP".to_string()
-                        } else {
-                            output_remover.replace_all(
-                                &x.1.to_value().split_whitespace().join(" "),
-                                &remover_vals,
-                            )
-                        };
-                        result_state.prev_message.insert(x.0.clone(), x.1.clone());
-                        ret
-                    }
-                    _ => output_remover
-                        .replace_all(&x.1.to_value().split_whitespace().join(" "), &remover_vals),
-                }
-            }))?;
-        }
-    }
-
-    if !output_writer.display_flag {
-        match &mut output_writer.result_writer {
-            ResultWriter::Csv(csv_writer) => csv_writer.flush()?,
-            ResultWriter::Json { writer, .. } => writer.flush()?,
-        }
-    }
-    if json_output_flag && stored_static.output_path.is_none() {
-        println!()
-    }
-    Ok(())
-}
-
-/// Folds a batch of detections into `result_state`: records the detection timestamps and the
-/// IDs of detected records, and (unless no-summary is set) updates the per-level counts by
-/// date, computer, and rule, plus the rule author statistics used by the results summary.
-fn calc_statistic_info(
-    detect_infos: &[DetectInfo],
-    duplicate_indices: &HashSet<usize>,
-    result_state: &mut ResultOutputState,
-    stored_static: &StoredStatic,
-) {
-    let output_option = stored_static.output_option.as_ref().unwrap();
-    for (i, detect_info) in detect_infos.iter().enumerate() {
-        if duplicate_indices.contains(&i) {
-            continue;
-        }
-        result_state
-            .timestamps
-            .push(detect_info.detected_time.timestamp());
-        match &detect_info.agg_result {
-            None => {
-                result_state
-                    .detected_record_idset
-                    .insert(CompactString::from(format!(
-                        "{}_{}",
-                        detect_info.detected_time, detect_info.eventid
-                    )));
-            }
-            Some(agg_result) => {
-                agg_result.agg_record_time_info.iter().for_each(|a| {
-                    result_state
-                        .detected_record_idset
-                        .insert(CompactString::from(format!("{}_{}", a.time, a.event_id)));
-                });
-            }
-        }
-        if !output_option.no_summary {
-            let level_index = detect_info.level.index();
-            let author_list = extract_author_name(&detect_info.ruleauthor);
-            let author_str = author_list.iter().join(", ");
-            result_state.detect_rule_authors.insert(
-                detect_info.rule_path.to_owned(),
-                author_str.to_string().into(),
-            );
-
-            if author_str != "-"
-                && !result_state
-                    .detected_rule_files
-                    .contains(&detect_info.rule_path)
-            {
-                result_state
-                    .detected_rule_files
-                    .insert(detect_info.rule_path.to_owned());
-                for author in author_list.iter() {
-                    *result_state
-                        .rule_author_counter
-                        .entry(CompactString::from(author))
-                        .or_insert(0) += 1;
-                }
-            }
-            if !result_state.detected_rule_ids.contains(&detect_info.ruleid) {
-                result_state
-                    .detected_rule_ids
-                    .insert(detect_info.ruleid.to_owned());
-                result_state.unique_detect_counts_by_level[level_index] += 1;
-            }
-            let computer_names = match &detect_info.agg_result {
-                None => vec![detect_info.computername.clone()],
-                Some(agg) => agg
-                    .agg_record_time_info
-                    .iter()
-                    .map(|a| CompactString::from(a.computer.clone()))
-                    .collect::<std::collections::HashSet<_>>() // Convert to HashSet to remove duplicates
-                    .into_iter()
-                    .sorted()
-                    .collect(),
-            };
-            for computername in &computer_names {
-                let computer_rule_check_key =
-                    CompactString::from(format!("{}|{}", computername, &detect_info.rule_path));
-                if !result_state
-                    .detected_computer_and_rule_names
-                    .contains(&computer_rule_check_key)
-                {
-                    result_state
-                        .detected_computer_and_rule_names
-                        .insert(computer_rule_check_key);
-                    countup_aggregation(
-                        &mut result_state.detect_counts_by_computer_and_level,
-                        &detect_info.level,
-                        computername,
-                    );
-                }
-            }
-            result_state.rule_title_path_map.insert(
-                detect_info.ruletitle.to_owned(),
-                detect_info.rule_path.to_owned(),
-            );
-
-            countup_aggregation(
-                &mut result_state.detect_counts_by_date_and_level,
-                &detect_info.level,
-                &format_time(
-                    &detect_info.detected_time,
-                    true,
-                    &output_option.time_format_options,
-                ),
-            );
-            countup_aggregation(
-                &mut result_state.detect_counts_by_rule_and_level,
-                &detect_info.level,
-                &detect_info.ruletitle,
-            );
-            let level_index = detect_info.level.index();
-            result_state.total_detect_counts_by_level[level_index] += 1;
-        }
-    }
-}
-
-/// Prints everything that follows the timeline itself: the rule author table, the detection
-/// frequency timeline (if requested), and the results summary (event counts, detection counts
-/// per level/date/computer/rule). Also accumulates the same content for the HTML report if enabled.
-pub fn output_result_summary(
-    stored_static: &StoredStatic,
-    output_writer: &mut OutputWriter,
-    result_state: &ResultOutputState,
-) {
-    if output_writer.display_flag {
-        println!();
-    }
-
-    let terminal_width = match terminal_size() {
-        Some((Width(w), _)) => w as usize,
-        None => 100,
-    };
-
-    let output_option = stored_static.output_option.as_ref().unwrap();
-    if !output_option.no_summary && !result_state.rule_author_counter.is_empty() {
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(
-                Some(Color::Rgb(0, 255, 0)),
-                stored_static.common_options.no_color,
-            ),
-            "Rule Authors:",
-            false,
-        )
-        .ok();
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(None, stored_static.common_options.no_color),
-            " ",
-            true,
-        )
-        .ok();
-
-        println!();
-        let table_column_num = if terminal_width <= 105 {
-            2
-        } else if terminal_width < 140 {
-            3
-        } else if terminal_width < 175 {
-            4
-        } else if terminal_width <= 210 {
-            5
-        } else {
-            6
-        };
-        output_detected_rule_authors(&result_state.rule_author_counter, table_column_num);
-    }
-
-    println!();
-    if output_option.visualize_timeline {
-        _print_timeline_hist(&result_state.timestamps, terminal_width, 3);
-        println!();
-    }
-
-    let mut html_output_stock = Nested::<String>::new();
-    if !output_option.no_summary {
-        output_writer.disp_wtr_buf.clear();
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(
-                Some(Color::Rgb(0, 255, 0)),
-                stored_static.common_options.no_color,
-            ),
-            "Results Summary:",
-            false,
-        )
-        .ok();
-
-        if let Some(timeline_start_time) = result_state.timeline_start_time {
-            output_and_data_stack_for_html(
-                &format!(
-                    "First timestamp: {}",
-                    utils::format_time(
-                        &timeline_start_time,
-                        false,
-                        &stored_static
-                            .output_option
-                            .as_ref()
-                            .unwrap()
-                            .time_format_options
-                    )
-                ),
-                "Results Summary {#results_summary}",
-                &stored_static.html_report_flag,
-            );
-        }
-        if let Some(timeline_end_time) = result_state.timeline_end_time {
-            output_and_data_stack_for_html(
-                &format!(
-                    "Last timestamp: {}",
-                    utils::format_time(
-                        &timeline_end_time,
-                        false,
-                        &stored_static
-                            .output_option
-                            .as_ref()
-                            .unwrap()
-                            .time_format_options
-                    )
-                ),
-                "Results Summary {#results_summary}",
-                &stored_static.html_report_flag,
-            );
-            println!();
-        }
-        if let Some(detect_starttime) = result_state.detect_starttime {
-            output_and_data_stack_for_html(
-                &format!(
-                    "First detection: {}",
-                    utils::format_time(
-                        &detect_starttime,
-                        false,
-                        &stored_static
-                            .output_option
-                            .as_ref()
-                            .unwrap()
-                            .time_format_options
-                    )
-                ),
-                "Results Summary {#results_summary}",
-                &stored_static.html_report_flag,
-            );
-        }
-        if let Some(detect_endtime) = result_state.detect_endtime {
-            output_and_data_stack_for_html(
-                &format!(
-                    "Last detection: {}",
-                    utils::format_time(
-                        &detect_endtime,
-                        false,
-                        &stored_static
-                            .output_option
-                            .as_ref()
-                            .unwrap()
-                            .time_format_options
-                    )
-                ),
-                "Results Summary {#results_summary}",
-                &stored_static.html_report_flag,
-            );
-            println!();
-        }
-
-        let reduced_record_cnt: u128 =
-            result_state.record_cnt - result_state.detected_record_idset.len() as u128;
-        let reduced_percent = if result_state.record_cnt == 0 {
-            0 as f64
-        } else {
-            (reduced_record_cnt as f64) / (result_state.record_cnt as f64) * 100.0
-        };
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(
-                Some(Color::Rgb(255, 255, 0)),
-                stored_static.common_options.no_color,
-            ),
-            "Events with hits",
-            false,
-        )
-        .ok();
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(None, stored_static.common_options.no_color),
-            " / ",
-            false,
-        )
-        .ok();
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(
-                Some(Color::Rgb(0, 255, 255)),
-                stored_static.common_options.no_color,
-            ),
-            "Total events",
-            false,
-        )
-        .ok();
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(None, stored_static.common_options.no_color),
-            ": ",
-            false,
-        )
-        .ok();
-        let saved_alerts_output =
-            (result_state.record_cnt - reduced_record_cnt).to_formatted_string(&Locale::en);
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(
-                Some(Color::Rgb(255, 255, 0)),
-                stored_static.common_options.no_color,
-            ),
-            &saved_alerts_output,
-            false,
-        )
-        .ok();
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(None, stored_static.common_options.no_color),
-            " / ",
-            false,
-        )
-        .ok();
-
-        let all_record_output = result_state.record_cnt.to_formatted_string(&Locale::en);
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(
-                Some(Color::Rgb(0, 255, 255)),
-                stored_static.common_options.no_color,
-            ),
-            &all_record_output,
-            false,
-        )
-        .ok();
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(None, stored_static.common_options.no_color),
-            " (",
-            false,
-        )
-        .ok();
-        let reduction_output = format!(
-            "Data reduction: {} events ({:.2}%)",
-            reduced_record_cnt.to_formatted_string(&Locale::en),
-            reduced_percent
-        );
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(
-                Some(Color::Rgb(0, 255, 0)),
-                stored_static.common_options.no_color,
-            ),
-            &reduction_output,
-            false,
-        )
-        .ok();
-
-        write_color_buffer(
-            &output_writer.display_writer,
-            get_writable_color(None, stored_static.common_options.no_color),
-            ")",
-            true,
-        )
-        .ok();
-        if stored_static.enable_recover_records {
-            write_color_buffer(
-                &output_writer.display_writer,
-                get_writable_color(
-                    Some(Color::Rgb(0, 255, 255)),
-                    stored_static.common_options.no_color,
-                ),
-                "Recovered records",
-                false,
-            )
-            .ok();
-            write_color_buffer(
-                &output_writer.display_writer,
-                get_writable_color(None, stored_static.common_options.no_color),
-                ": ",
-                false,
-            )
-            .ok();
-            let recovered_record_output = result_state
-                .recover_record_cnt
-                .to_formatted_string(&Locale::en);
-            write_color_buffer(
-                &output_writer.display_writer,
-                get_writable_color(
-                    Some(Color::Rgb(0, 255, 255)),
-                    stored_static.common_options.no_color,
-                ),
-                &recovered_record_output,
-                true,
-            )
-            .ok();
-        }
-        println!();
-
-        if stored_static.html_report_flag {
-            html_output_stock.push(format!("- Events with hits: {}", &saved_alerts_output));
-            html_output_stock.push(format!("- Total events analyzed: {}", &all_record_output));
-            html_output_stock.push(format!("- {reduction_output}"));
-            html_output_stock.push(format!(
-                "- Recovered events analyzed: {}",
-                &result_state
-                    .recover_record_cnt
-                    .to_formatted_string(&Locale::en)
-            ));
-        }
-
-        let color_map = create_output_color_map(stored_static.common_options.no_color);
-        _print_unique_results(
-            &result_state.total_detect_counts_by_level,
-            &result_state.unique_detect_counts_by_level,
-            (
-                CompactString::from("Total | Unique"),
-                CompactString::from("detections"),
-            ),
-            &color_map,
-            &mut html_output_stock,
-            stored_static.html_report_flag,
-        );
-        println!();
-        if let Some(timeline_start_time) = result_state.timeline_start_time {
-            let ts = format!(
-                "First timestamp: {}",
-                format_time(
-                    &timeline_start_time,
-                    false,
-                    &stored_static
-                        .output_option
-                        .as_ref()
-                        .unwrap()
-                        .time_format_options
-                )
-            );
-            write_color_buffer(
-                &BufferWriter::stdout(ColorChoice::Always),
-                None,
-                ts.as_str(),
-                true,
-            )
-            .ok();
-        }
-        if let Some(timeline_end_time) = result_state.timeline_end_time {
-            let ts = format!(
-                "Last timestamp: {}",
-                format_time(
-                    &timeline_end_time,
-                    false,
-                    &stored_static
-                        .output_option
-                        .as_ref()
-                        .unwrap()
-                        .time_format_options
-                )
-            );
-            write_color_buffer(
-                &BufferWriter::stdout(ColorChoice::Always),
-                None,
-                ts.as_str(),
-                true,
-            )
-            .ok();
-            println!();
-        }
-        if let Some(detect_starttime) = result_state.detect_starttime {
-            let ts = format!(
-                "First detection: {}",
-                format_time(
-                    &detect_starttime,
-                    false,
-                    &stored_static
-                        .output_option
-                        .as_ref()
-                        .unwrap()
-                        .time_format_options
-                )
-            );
-            write_color_buffer(
-                &BufferWriter::stdout(ColorChoice::Always),
-                None,
-                ts.as_str(),
-                true,
-            )
-            .ok();
-        }
-        if let Some(detect_endtime) = result_state.detect_endtime {
-            let ts = format!(
-                "Last detection: {}",
-                format_time(
-                    &detect_endtime,
-                    false,
-                    &stored_static
-                        .output_option
-                        .as_ref()
-                        .unwrap()
-                        .time_format_options
-                )
-            );
-            write_color_buffer(
-                &BufferWriter::stdout(ColorChoice::Always),
-                None,
-                ts.as_str(),
-                true,
-            )
-            .ok();
-            println!();
-        }
-
-        _print_detection_summary_by_date(
-            &result_state.detect_counts_by_date_and_level,
-            &color_map,
-            &mut html_output_stock,
-            stored_static,
-        );
-        println!();
-        println!();
-        if stored_static.html_report_flag {
-            html_output_stock.push("");
-        }
-
-        _print_detection_summary_by_computer(
-            &result_state.detect_counts_by_computer_and_level,
-            &color_map,
-            &mut html_output_stock,
-            stored_static,
-        );
-        println!();
-        if stored_static.html_report_flag {
-            html_output_stock.push("");
-        }
-
-        _print_detection_summary_tables(
-            &result_state.detect_counts_by_rule_and_level,
-            &color_map,
-            (
-                &result_state.rule_title_path_map,
-                &result_state.detect_rule_authors,
-            ),
-            &mut html_output_stock,
-            stored_static,
-            cmp::min((terminal_width / 2) - 15, 200),
-        );
-        println!();
-        if stored_static.html_report_flag {
-            html_output_stock.push("");
-        }
-    }
-    if stored_static.html_report_flag {
-        _output_html_computer_by_mitre_attck(&mut html_output_stock);
-        htmlreport::add_md_data("Results Summary {#results_summary}", html_output_stock);
-    }
 }
 
 /// Sorts detections by detected time, level, event ID, rule path, computer name, record ID,
@@ -1189,590 +415,6 @@ pub fn get_duplicate_indices(detect_infos: &mut [DetectInfo]) -> HashSet<usize> 
     filtered_detect_infos
 }
 
-fn _get_table_color(
-    color_map: &HashMap<LEVEL, Colors>,
-    level: &LEVEL,
-) -> Option<comfy_table::Color> {
-    let mut color = None;
-    if let Some(c) = color_map.get(level) {
-        color = Some(c.table_color);
-    }
-    color
-}
-
-/// Prints the detection frequency timeline (a sparkline histogram of detection timestamps with
-/// time markers) to stdout. Requires at least 5 events to render.
-fn _print_timeline_hist(timestamps: &[i64], length: usize, side_margin_size: usize) {
-    if timestamps.is_empty() {
-        return;
-    }
-
-    let buf_wtr = BufferWriter::stdout(ColorChoice::Always);
-    let mut wtr = buf_wtr.buffer();
-    wtr.set_color(ColorSpec::new().set_fg(None)).ok();
-
-    if timestamps.len() < 5 {
-        writeln!(
-            wtr,
-            "Detection Frequency Timeline could not be displayed as there needs to be more than 5 events.",
-        )
-        .ok();
-        buf_wtr.print(&wtr).ok();
-        return;
-    }
-
-    let title = "Detection Frequency Timeline";
-    let header_row_space = (length - title.len()) / 2;
-    writeln!(wtr, "{}{}", " ".repeat(header_row_space), title).ok();
-    println!();
-
-    let timestamp_marker_max = if timestamps.len() < 2 {
-        0
-    } else {
-        timestamps.len() - 2
-    };
-    let marker_num = min(timestamp_marker_max, 18);
-
-    let (header_raw, footer_raw) =
-        build_time_markers(timestamps, marker_num, length - (side_margin_size * 2));
-    let sparkline = build_sparkline(timestamps, length - (side_margin_size * 2), 5_usize);
-    for header_str in header_raw.lines() {
-        writeln!(wtr, "{}{}", " ".repeat(side_margin_size - 1), header_str).ok();
-    }
-    for line in sparkline.lines() {
-        writeln!(wtr, "{}{}", " ".repeat(side_margin_size - 1), line).ok();
-    }
-    for footer_str in footer_raw.lines() {
-        writeln!(wtr, "{}{}", " ".repeat(side_margin_size - 1), footer_str).ok();
-    }
-
-    buf_wtr.print(&wtr).ok();
-}
-
-/// Increments the counter for `entry_key` in the inner map for `level`, falling back to the
-/// UNDEFINED level's map if the level has no entry.
-fn countup_aggregation(
-    count_map: &mut HashMap<LEVEL, HashMap<CompactString, i128>>,
-    level: &LEVEL,
-    entry_key: &str,
-) {
-    let mut detect_counts_by_rules = count_map
-        .get(level)
-        .unwrap_or_else(|| count_map.get(&LEVEL::UNDEFINED).unwrap())
-        .to_owned();
-    *detect_counts_by_rules.entry(entry_key.into()).or_insert(0) += 1;
-    count_map.insert(level.clone(), detect_counts_by_rules);
-}
-
-/// Column position within a display cell, which determines where padding spaces are added:
-/// First: |<str> |
-/// Last: | <str>|
-/// Other: | <str> |
-enum ColPos {
-    First,
-    Last,
-    Other,
-}
-
-/// Writes one record (or, when `header` is true, the column header row) to the terminal in the
-/// "·"-separated display format, coloring the timestamp/level/rule-title fields by detection
-/// level and the field names/values inside the details sections individually.
-fn _get_serialized_disp_output(
-    display_writer: &BufferWriter,
-    data: &[(CompactString, Profile)],
-    header: bool,
-    (output_remover, remover_vals): (&AhoCorasick, &[&str]),
-    no_color: bool,
-    level_color: Option<Color>,
-) {
-    let data_length = data.len();
-    let mut ret = Nested::<String>::new();
-    if header {
-        for (i, d) in data.iter().enumerate() {
-            if i == 0 {
-                ret.push(_format_cellpos(&d.0, ColPos::First))
-            } else if i == data_length - 1 {
-                ret.push(_format_cellpos(&d.0, ColPos::Last))
-            } else {
-                ret.push(_format_cellpos(&d.0, ColPos::Other))
-            }
-        }
-        let mut disp_serializer = WriterBuilder::new()
-            .double_quote(false)
-            .quote_style(QuoteStyle::Never)
-            .delimiter(b'|')
-            .has_headers(false)
-            .from_writer(vec![]);
-
-        disp_serializer
-            .write_record(ret.iter().collect::<Vec<_>>())
-            .ok();
-
-        write_color_buffer(
-            display_writer,
-            get_writable_color(None, no_color),
-            // The serializer above uses '|' as its delimiter; show those separators as '·' and
-            // then restore any '🦅' placeholders back to '|'. Historically cell values had
-            // literal '|' replaced with '🦅' before serialization so this delimiter
-            // substitution would not clobber them; nothing currently inserts that placeholder
-            // on this path, so the second replace is defensive/vestigial.
-            &String::from_utf8(disp_serializer.into_inner().unwrap_or_default())
-                .unwrap_or_default()
-                .replace('|', "·")
-                .replace('🦅', "|"),
-            false,
-        )
-        .ok();
-    } else {
-        for (i, d) in data.iter().enumerate() {
-            let col_pos = if i == 0 {
-                ColPos::First
-            } else if i == data_length - 1 {
-                ColPos::Last
-            } else {
-                ColPos::Other
-            };
-            let display_contents = _format_cellpos(
-                &output_remover
-                    .replace_all(&d.1.to_value().split_whitespace().join(" "), remover_vals)
-                    .split_ascii_whitespace()
-                    .join(" "),
-                col_pos,
-            );
-            let output_color_and_contents = match d.1 {
-                Profile::Timestamp(_) | Profile::Level(_) | Profile::RuleTitle(_) => {
-                    vec![vec![(
-                        display_contents,
-                        get_writable_color(level_color, no_color),
-                    )]]
-                }
-                Profile::AllFieldInfo(_) | Profile::Details(_) | Profile::ExtraFieldInfo(_) => {
-                    let mut output_str_char_pair = vec![];
-                    for c in display_contents.split('¦') {
-                        if let Some((field, val)) = c.split_once(':') {
-                            let mut field_val_col_pair = vec![];
-                            field_val_col_pair.push((
-                                format!(" {}: ", field.trim()),
-                                get_writable_color(Some(Color::Rgb(255, 158, 61)), no_color),
-                            ));
-
-                            field_val_col_pair.push((
-                                format!(
-                                    "{} ",
-                                    output_remover
-                                        .replace_all(
-                                            &val.split_whitespace().join(" "),
-                                            remover_vals,
-                                        )
-                                        .split_ascii_whitespace()
-                                        .join(" ")
-                                ),
-                                get_writable_color(Some(Color::Rgb(0, 255, 255)), no_color),
-                            ));
-                            output_str_char_pair.push(field_val_col_pair);
-                        }
-                    }
-                    if output_str_char_pair.is_empty() {
-                        vec![vec![(display_contents, None)]]
-                    } else {
-                        output_str_char_pair
-                    }
-                }
-                _ => {
-                    vec![vec![(display_contents, None)]]
-                }
-            };
-
-            let col_cnt = output_color_and_contents.len();
-            for (field_idx, col_contents) in output_color_and_contents.iter().enumerate() {
-                for (c, color) in col_contents {
-                    write_color_buffer(display_writer, *color, c, false).ok();
-                }
-                if field_idx != col_cnt - 1 {
-                    write_color_buffer(display_writer, None, "¦", false).ok();
-                }
-            }
-
-            if i != data_length - 1 {
-                write_color_buffer(
-                    display_writer,
-                    get_writable_color(Some(Color::Rgb(255, 158, 61)), no_color),
-                    "·",
-                    false,
-                )
-                .ok();
-            } else {
-                // Line break after the last element of one record (plus a blank separator line).
-                println!();
-                println!();
-            }
-        }
-    }
-}
-
-/// Pads a cell value with spaces according to its column position in the display output.
-fn _format_cellpos(colval: &str, column: ColPos) -> String {
-    match column {
-        ColPos::First => format!("{colval} "),
-        ColPos::Last => format!(" {colval}"),
-        ColPos::Other => format!(" {colval} "),
-    }
-}
-
-/// Prints the total and unique detection counts (overall and per level, with percentages) to
-/// stdout, and accumulates the same information for the HTML report when enabled.
-fn _print_unique_results(
-    counts_by_level: &[u128],
-    unique_counts_by_level: &[u128],
-    head_and_tail_word: (CompactString, CompactString),
-    color_map: &HashMap<LEVEL, Colors>,
-    html_output_stock: &mut Nested<String>,
-    html_output_flag: bool,
-) {
-    // The counts are stored in ascending level order, but levels are displayed from highest to
-    // lowest, so iterate over them in reverse.
-    let mut counts_by_level_rev = counts_by_level.iter().rev();
-    let mut unique_counts_by_level_rev = unique_counts_by_level.iter().rev();
-
-    let total_count = counts_by_level.iter().sum::<u128>();
-    let unique_total_count = unique_counts_by_level.iter().sum::<u128>();
-    // Output the totals across all levels first.
-    write_color_buffer(
-        &BufferWriter::stdout(ColorChoice::Always),
-        None,
-        &format!(
-            "{} {}: {} | {}",
-            head_and_tail_word.0,
-            head_and_tail_word.1,
-            total_count.to_formatted_string(&Locale::en),
-            unique_total_count.to_formatted_string(&Locale::en)
-        ),
-        true,
-    )
-    .ok();
-
-    let mut total_detect_md = vec!["- Total detections:".to_string()];
-    let mut unique_detect_md = vec!["- Unique detections:".to_string()];
-
-    for level_name in LEVEL::iter().rev() {
-        if level_name == LEVEL::UNDEFINED {
-            continue;
-        }
-        let count_by_level = *counts_by_level_rev.next().unwrap();
-        let unique_count_by_level = *unique_counts_by_level_rev.next().unwrap();
-        let percent = if total_count == 0 {
-            0 as f64
-        } else {
-            (count_by_level as f64) / (total_count as f64) * 100.0
-        };
-        let unique_percent = if unique_total_count == 0 {
-            0 as f64
-        } else {
-            (unique_count_by_level as f64) / (unique_total_count as f64) * 100.0
-        };
-        if html_output_flag {
-            total_detect_md.push(format!(
-                "    - {}: {} ({:.2}%)",
-                level_name.to_full(),
-                count_by_level.to_formatted_string(&Locale::en),
-                percent
-            ));
-            unique_detect_md.push(format!(
-                "    - {}: {} ({:.2}%)",
-                level_name.to_full(),
-                unique_count_by_level.to_formatted_string(&Locale::en),
-                unique_percent
-            ));
-        }
-        let output_raw_str = format!(
-            "{} {} {}: {} ({:.2}%) | {} ({:.2}%)",
-            head_and_tail_word.0,
-            level_name.to_full(),
-            head_and_tail_word.1,
-            count_by_level.to_formatted_string(&Locale::en),
-            percent,
-            unique_count_by_level.to_formatted_string(&Locale::en),
-            unique_percent
-        );
-        write_color_buffer(
-            &BufferWriter::stdout(ColorChoice::Always),
-            _get_output_color(color_map, &level_name),
-            &output_raw_str,
-            true,
-        )
-        .ok();
-    }
-    if html_output_flag {
-        html_output_stock.extend(total_detect_md.iter());
-        html_output_stock.extend(unique_detect_md.iter());
-    }
-}
-
-/// Output the date with the highest detection count for each level.
-fn _print_detection_summary_by_date(
-    detect_counts_by_date: &HashMap<LEVEL, HashMap<CompactString, i128>>,
-    color_map: &HashMap<LEVEL, Colors>,
-    html_output_stock: &mut Nested<String>,
-    stored_static: &StoredStatic,
-) {
-    let buf_wtr = BufferWriter::stdout(ColorChoice::Always);
-    let mut wtr = buf_wtr.buffer();
-    wtr.set_color(ColorSpec::new().set_fg(None)).ok();
-    let output_header = "Dates with most total detections:";
-    write_color_buffer(&buf_wtr, None, output_header, true).ok();
-
-    if stored_static.html_report_flag {
-        html_output_stock.push(format!("- {output_header}"));
-    }
-    for (i, level) in LEVEL::iter().rev().enumerate() {
-        if level == LEVEL::UNDEFINED {
-            continue;
-        }
-        // An inner map was inserted for every level when ResultOutputState was initialized, so the
-        // lookup is guaranteed to be Some and unwrap is called directly.
-        let detections_by_day = detect_counts_by_date.get(&level).unwrap();
-        let mut max_detect_str = CompactString::default();
-        let mut max_date: Option<&CompactString> = None;
-        let mut tmp_cnt: i128 = 0;
-        let mut exist_max_data = false;
-        for (date, cnt) in detections_by_day {
-            // On a tie, pick the lexicographically-smallest date string so the
-            // choice is deterministic rather than HashMap-iteration-dependent.
-            // (The key is the already-formatted date, so this is lexical, not
-            // necessarily chronological under non-ISO date formats — fine, since
-            // only determinism matters here.)
-            if *cnt > tmp_cnt || (*cnt == tmp_cnt && exist_max_data && Some(date) < max_date) {
-                exist_max_data = true;
-                max_date = Some(date);
-                tmp_cnt = *cnt;
-            }
-        }
-        if let Some(date) = max_date {
-            max_detect_str =
-                format!("{} ({})", date, tmp_cnt.to_formatted_string(&Locale::en)).into();
-        }
-        wtr.set_color(ColorSpec::new().set_fg(_get_output_color(color_map, &level)))
-            .ok();
-        if !exist_max_data {
-            max_detect_str = "n/a".into();
-        }
-        let output_str = format!("{}: {}", level.to_full(), &max_detect_str);
-        write!(wtr, "{output_str}").ok();
-        // Print a ", " separator after every level except the last displayed one (count() - 2
-        // because UNDEFINED, the final item of the reversed iteration, is skipped).
-        if i != LEVEL::iter().count() - 2 {
-            wtr.set_color(ColorSpec::new().set_fg(None)).ok();
-            write!(wtr, ", ").ok();
-        }
-        if stored_static.html_report_flag {
-            html_output_stock.push(format!("    - {output_str}"));
-        }
-    }
-    buf_wtr.print(&wtr).ok();
-}
-
-/// Output the top 5 computers with the most unique detections for each level (the HTML report
-/// gets the full list).
-fn _print_detection_summary_by_computer(
-    detect_counts_by_computer: &HashMap<LEVEL, HashMap<CompactString, i128>>,
-    color_map: &HashMap<LEVEL, Colors>,
-    html_output_stock: &mut Nested<String>,
-    stored_static: &StoredStatic,
-) {
-    let buf_wtr = BufferWriter::stdout(ColorChoice::Always);
-    let mut wtr = buf_wtr.buffer();
-    wtr.set_color(ColorSpec::new().set_fg(None)).ok();
-
-    writeln!(wtr, "Top 5 computers with most unique detections:").ok();
-    for level in LEVEL::iter().rev() {
-        if level == LEVEL::UNDEFINED {
-            continue;
-        }
-        // An inner map was inserted for every level when ResultOutputState was initialized, so the
-        // lookup is guaranteed to be Some and unwrap is called directly.
-        let detections_by_computer = detect_counts_by_computer.get(&level).unwrap();
-        let mut result_vec = Nested::<String>::new();
-        // Exclude entries where the computer name is "-" from the aggregation.
-        let mut sorted_detections: Vec<(&CompactString, &i128)> = detections_by_computer
-            .iter()
-            .filter(|a| a.0.as_str() != "-")
-            .collect();
-
-        // Sort by count descending, then by name ascending so equal-count
-        // entries have a deterministic (rather than HashMap-iteration-dependent)
-        // order.
-        sorted_detections.sort_by(|a, b| b.1.cmp(a.1).then_with(|| a.0.cmp(b.0)));
-
-        // For HTML output, display all computer names.
-        if stored_static.html_report_flag {
-            html_output_stock.push(format!(
-                "### Computers with most unique {} detections: {{#computers_with_most_unique_{}_detections}}",
-                level.to_full(),
-                level.to_full(),
-            ));
-            for x in sorted_detections.iter() {
-                html_output_stock.push(format!(
-                    "- {} ({})",
-                    html_escape_value(x.0),
-                    x.1.to_formatted_string(&Locale::en)
-                ));
-            }
-            html_output_stock.push("");
-        }
-        for x in sorted_detections.iter().take(5) {
-            result_vec.push(format!(
-                "{} ({})",
-                x.0,
-                x.1.to_formatted_string(&Locale::en)
-            ));
-        }
-        let result_str = if result_vec.is_empty() {
-            "n/a".to_string()
-        } else {
-            result_vec.iter().collect::<Vec<_>>().join(", ")
-        };
-
-        wtr.set_color(ColorSpec::new().set_fg(_get_output_color(color_map, &level)))
-            .ok();
-        writeln!(wtr, "{}: {}", level.to_full(), &result_str).ok();
-    }
-    buf_wtr.print(&wtr).ok();
-}
-
-/// Output the rules with the most detections for each level in table format (top 5 per level;
-/// the HTML report gets the full list). Rule titles longer than `limit_num` characters are
-/// truncated with "...".
-fn _print_detection_summary_tables(
-    detect_counts_by_rule_and_level: &HashMap<LEVEL, HashMap<CompactString, i128>>,
-    color_map: &HashMap<LEVEL, Colors>,
-    (rule_title_path_map, rule_detect_author_map): (
-        &HashMap<CompactString, CompactString>,
-        &HashMap<CompactString, CompactString>,
-    ),
-    html_output_stock: &mut Nested<String>,
-    stored_static: &StoredStatic,
-    limit_num: usize,
-) {
-    let buf_wtr = BufferWriter::stdout(ColorChoice::Always);
-    let mut wtr = buf_wtr.buffer();
-    wtr.set_color(ColorSpec::new().set_fg(None)).ok();
-    let mut output = vec![];
-    let mut col_color = vec![];
-    for level in LEVEL::iter().rev() {
-        if level == LEVEL::UNDEFINED {
-            continue;
-        }
-
-        let mut col_output: Nested<String> = Nested::<String>::new();
-        col_output.push(format!("Top {} alerts:", level.to_full()));
-
-        col_color.push(_get_table_color(color_map, &level));
-
-        // An inner map was inserted for every level when ResultOutputState was initialized, so the
-        // lookup is guaranteed to be Some and unwrap is called directly.
-        let detections_by_computer = detect_counts_by_rule_and_level.get(&level).unwrap();
-        let mut sorted_detections: Vec<(&CompactString, &i128)> =
-            detections_by_computer.iter().collect();
-
-        // Sort by count descending, then by name ascending so equal-count
-        // entries have a deterministic (rather than HashMap-iteration-dependent)
-        // order.
-        sorted_detections.sort_by(|a, b| b.1.cmp(a.1).then_with(|| a.0.cmp(b.0)));
-
-        // For HTML output, list every rule rather than only the top five.
-        if stored_static.html_report_flag {
-            html_output_stock.push(format!(
-                "### All {} alerts: {{#all_{}_alerts}}",
-                level.to_full(),
-                level.to_full()
-            ));
-            let rule_path = stored_static.output_option.as_ref().unwrap().rules.clone();
-            let rule_encoded =
-                check_setting_path(&CURRENT_EXE_PATH.to_path_buf(), "encoded_rules.yml", true)
-                    .unwrap();
-            let is_encoded_rule = rule_encoded.exists() && rule_encoded.to_path_buf() == rule_path;
-            for x in sorted_detections.iter() {
-                let not_found_str = CompactString::from_str("<Not Found Path>").unwrap();
-                let rule_path = rule_title_path_map.get(x.0).unwrap_or(&not_found_str);
-                if is_encoded_rule {
-                    html_output_stock.push(format!(
-                        "- {} ({}) - {}",
-                        html_escape_value(x.0),
-                        x.1.to_formatted_string(&Locale::en),
-                        html_escape_value(
-                            rule_detect_author_map
-                                .get(rule_path)
-                                .unwrap_or(&not_found_str)
-                        )
-                    ));
-                } else {
-                    html_output_stock.push(format!(
-                        "- [{}]({}) ({}) - {}",
-                        html_escape_value(x.0),
-                        &rule_path.replace('\\', "/"),
-                        x.1.to_formatted_string(&Locale::en),
-                        html_escape_value(
-                            rule_detect_author_map
-                                .get(rule_path)
-                                .unwrap_or(&not_found_str)
-                        )
-                    ));
-                }
-            }
-            html_output_stock.push("");
-        }
-
-        let take_cnt = 5;
-        for x in sorted_detections.iter().take(take_cnt) {
-            let output_title = if x.0.len() > limit_num - 3 {
-                format!("{}...", &x.0[..(limit_num - 3)])
-            } else {
-                x.0.to_string()
-            };
-            col_output.push(format!(
-                "{output_title} ({})",
-                x.1.to_formatted_string(&Locale::en)
-            ));
-        }
-        let na_cnt = if sorted_detections.len() > take_cnt {
-            0
-        } else {
-            take_cnt - sorted_detections.len()
-        };
-        for _x in 0..na_cnt {
-            col_output.push("n/a");
-        }
-        output.push(col_output);
-    }
-
-    let mut tb = Table::new();
-    tb.load_preset(UTF8_FULL)
-        .apply_modifier(UTF8_ROUND_CORNERS)
-        .set_style(TableComponent::VerticalLines, ' ');
-    let horizontal_line_char = tb.style(TableComponent::HorizontalLines).unwrap();
-    let top_border_char = tb.style(TableComponent::TopBorder).unwrap();
-    for x in 0..output.len() / 2 {
-        tb.add_row(vec![
-            Cell::new(&output[2 * x][0]).fg(col_color[2 * x].unwrap_or(comfy_table::Color::Reset)),
-            Cell::new(&output[2 * x + 1][0])
-                .fg(col_color[2 * x + 1].unwrap_or(comfy_table::Color::Reset)),
-        ])
-        .set_style(TableComponent::MiddleIntersections, horizontal_line_char)
-        .set_style(TableComponent::TopBorderIntersections, top_border_char)
-        .set_style(
-            TableComponent::BottomBorderIntersections,
-            horizontal_line_char,
-        );
-        tb.add_row(vec![
-            Cell::new(output[2 * x].iter().skip(1).join("\n"))
-                .fg(col_color[2 * x].unwrap_or(comfy_table::Color::Reset)),
-            Cell::new(output[2 * x + 1].iter().skip(1).join("\n"))
-                .fg(col_color[2 * x + 1].unwrap_or(comfy_table::Color::Reset)),
-        ]);
-    }
-    println!("{tb}");
-}
-
 /// Converts the given datetime to epoch seconds. Unless UTC or ISO 8601 output was requested,
 /// the local UTC offset is added so the value reflects local wall-clock time.
 fn _get_timestamp(output_option: &TimeFormatOptions, time: &DateTime<Utc>) -> i64 {
@@ -1786,467 +428,6 @@ fn _get_timestamp(output_option: &TimeFormatOptions, time: &DateTime<Utc>) -> i6
             .local_minus_utc();
         offset_sec as i64 + time.with_timezone(&Local).timestamp()
     }
-}
-
-/// Splits the value into its elements for the profile members that are output as JSON arrays or
-/// objects: MitreTactics/MitreTags/OtherTags (": "-separated) and Details/AllFieldInfo/
-/// ExtraFieldInfo (" ¦ "-separated key-value pairs). Returns an empty Vec for everything else,
-/// and also for a details value that is a single element with no "key: value" structure.
-fn _get_json_vec(profile: &Profile, target_data: &String) -> Vec<String> {
-    match profile {
-        Profile::MitreTactics(_) | Profile::MitreTags(_) | Profile::OtherTags(_) => {
-            target_data.split(": ").map(|x| x.to_string()).collect()
-        }
-        Profile::Details(_) | Profile::AllFieldInfo(_) | Profile::ExtraFieldInfo(_) => {
-            let ret: Vec<String> = target_data.split(" ¦ ").map(|x| x.to_string()).collect();
-            if target_data == &ret[0] && !utils::contains_str(target_data, ": ") {
-                vec![]
-            } else {
-                ret
-            }
-        }
-        _ => vec![],
-    }
-}
-
-/// An ordered JSON value used to assemble one timeline record. Objects preserve key
-/// insertion order (via `IndexMap`) so the output keeps the rule/field order — unlike
-/// `serde_json::Map`, which sorts keys. Leaves are plain `serde_json::Value`s (scalars or
-/// arrays), so `serde_json` performs all escaping and number/bool serialization.
-enum JsonNode {
-    Leaf(serde_json::Value),
-    Object(IndexMap<String, JsonNode>),
-}
-
-impl Serialize for JsonNode {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        match self {
-            JsonNode::Leaf(value) => value.serialize(serializer),
-            // Serialize entries in insertion order (IndexMap iteration order), instead of the
-            // key-sorted order a `serde_json::Map` would produce.
-            JsonNode::Object(map) => {
-                let mut m = serializer.serialize_map(Some(map.len()))?;
-                for (k, v) in map {
-                    m.serialize_entry(k, v)?;
-                }
-                m.end()
-            }
-        }
-    }
-}
-
-/// Converts a field-value string to a JSON scalar, reproducing the previous type-guessing:
-/// an i64-parseable value becomes a JSON number, `true`/`false` becomes a boolean, and
-/// anything else becomes a string. Any real `\n`/`\r`/`\t` in the value are escaped by
-/// `serde_json` (as of #1849; previously they were `🛂n`/`🛂r`/`🛂t` placeholders rendered as the
-/// visible text `\\n`/`\\r`/`\\t`).
-fn json_scalar(value: &str) -> serde_json::Value {
-    if let Ok(i) = i64::from_str(value) {
-        serde_json::Value::from(i)
-    } else if let Ok(b) = bool::from_str(value) {
-        serde_json::Value::from(b)
-    } else {
-        serde_json::Value::String(close_unterminated_quote(value))
-    }
-}
-
-/// A JSON string leaf (no type-guessing) — used for array elements, which are always emitted
-/// as strings.
-fn json_string(value: &str) -> serde_json::Value {
-    serde_json::Value::String(close_unterminated_quote(value))
-}
-
-/// Reproduces the previous `_convert_valid_json_str` quirk where a value starting with a double
-/// quote but not ending with one had a closing quote appended (kept for byte-identical JSON data).
-fn close_unterminated_quote(value: &str) -> String {
-    let mut out = value.to_string();
-    if value.starts_with('"') && !value.ends_with('"') {
-        out.push('"');
-    }
-    out
-}
-
-/// Splits each "key: value" detail entry (on the first colon), trims both sides, and groups
-/// the values by key preserving first-seen key order. A key that appears more than once
-/// (e.g. Data[1]/Data[2]) collects multiple values, later emitted as a JSON array.
-fn group_details_raw(stock: &[CompactString]) -> IndexMap<String, Vec<String>> {
-    let mut map: IndexMap<String, Vec<String>> = IndexMap::new();
-    for contents in stock {
-        let (key, value) = contents.split_once(':').unwrap_or_default();
-        map.entry(key.trim().to_string())
-            .or_default()
-            .push(value.trim().to_string());
-    }
-    map
-}
-
-/// Serializes one record and returns the body *without* the surrounding object braces, so the
-/// caller's wrapping (`{\n...\n}` for JSON, `{ ... }` for JSONL) reproduces the full object.
-/// JSON uses a 4-space pretty formatter; JSONL is compact.
-fn serialize_record_body(record: IndexMap<String, JsonNode>, jsonl: bool) -> String {
-    if record.is_empty() {
-        return String::new();
-    }
-    let node = JsonNode::Object(record);
-    // Serializing this in-memory JsonNode (string keys, serde_json::Value leaves) to a String /
-    // Vec<u8> cannot fail; `expect` surfaces a would-be bug loudly instead of silently emitting
-    // empty/malformed JSON.
-    if jsonl {
-        let full = serde_json::to_string(&node).expect("JsonNode serialization is infallible");
-        full.strip_prefix('{')
-            .and_then(|s| s.strip_suffix('}'))
-            .unwrap_or(&full)
-            .to_string()
-    } else {
-        let mut buf = Vec::new();
-        let formatter = serde_json::ser::PrettyFormatter::with_indent(b"    ");
-        let mut ser = serde_json::Serializer::with_formatter(&mut buf, formatter);
-        node.serialize(&mut ser)
-            .expect("JsonNode serialization is infallible");
-        let full = String::from_utf8(buf).expect("serde_json output is valid UTF-8");
-        full.strip_prefix("{\n")
-            .and_then(|s| s.strip_suffix("\n}"))
-            .unwrap_or(&full)
-            .to_string()
-    }
-}
-
-/// Builds the JSON object body for one detection. Returns the body string (without the
-/// surrounding braces, which the caller adds) together with the updated previous-record field
-/// map used for duplicate-data suppression on the next record.
-pub fn output_json_str(
-    detect_info: &DetectInfo,
-    result_state: &mut ResultOutputState,
-    jsonl_output_flag: bool,
-    is_included_geo_ip: bool,
-    remove_duplicate_flag: bool,
-) -> (String, HashMap<CompactString, Profile>) {
-    let mut record: IndexMap<String, JsonNode> = IndexMap::new();
-    let mut target_ext_field = Vec::new();
-    let ext_field_map: HashMap<CompactString, Profile> =
-        HashMap::from_iter(detect_info.output_fields.to_owned());
-    let mut next_prev_message = result_state.prev_message.clone();
-    if remove_duplicate_flag {
-        for (field_name, profile) in detect_info.output_fields.iter() {
-            match profile {
-                Profile::Details(_) | Profile::AllFieldInfo(_) | Profile::ExtraFieldInfo(_) => {
-                    let details_key = match profile {
-                        Profile::Details(_) => "Details",
-                        Profile::AllFieldInfo(_) => "AllFieldInfo",
-                        Profile::ExtraFieldInfo(_) => "ExtraFieldInfo",
-                        _ => "",
-                    };
-
-                    let empty = vec![];
-                    let now = detect_info
-                        .details_convert_map
-                        .get(format!("#{details_key}").as_str())
-                        .unwrap_or(&empty);
-                    let prev = result_state
-                        .prev_details_convert_map
-                        .get(format!("#{details_key}").as_str())
-                        .unwrap_or(&empty);
-                    let dup_flag = (!profile.to_value().is_empty()
-                        && result_state
-                            .prev_message
-                            .get(field_name)
-                            .unwrap_or(&Profile::Literal("".into()))
-                            .to_value()
-                            == profile.to_value())
-                        || (!&now.is_empty() && !&prev.is_empty() && now == prev);
-                    if dup_flag {
-                        // Duplicate of the previous record: output the plain string "DUP"
-                        // instead of the value (Profile::Literal emits it as-is). The previous
-                        // message is intentionally NOT updated, so consecutive duplicates keep
-                        // being compared against the last non-duplicate value.
-                        target_ext_field.push((field_name.clone(), Profile::Literal("DUP".into())));
-                    } else {
-                        // Not a duplicate: remember this value for comparison with the next
-                        // record.
-                        next_prev_message.insert(field_name.clone(), profile.clone());
-                        target_ext_field.push((field_name.clone(), profile.clone()));
-                    }
-                }
-                _ => {
-                    target_ext_field.push((field_name.clone(), profile.clone()));
-                }
-            }
-        }
-    } else {
-        target_ext_field.clone_from(&detect_info.output_fields);
-    }
-    // GeoIP enrichment fields that are folded into the Details (and AllFieldInfo) objects of
-    // the JSON output instead of being emitted as top-level keys.
-    let key_add_to_details = [
-        "SrcASN",
-        "SrcCountry",
-        "SrcCity",
-        "TgtASN",
-        "TgtCountry",
-        "TgtCity",
-    ];
-
-    let valid_key_add_to_details: Vec<&str> = key_add_to_details
-        .iter()
-        .filter(|target_key| {
-            let target = ext_field_map.get(&CompactString::from(**target_key));
-            target.is_some() && target.unwrap().to_value() != "-"
-        })
-        .copied()
-        .collect();
-    for (key, profile) in target_ext_field.iter() {
-        let val = profile.to_value();
-        if !matches!(
-            profile,
-            Profile::AllFieldInfo(_) | Profile::ExtraFieldInfo(_)
-        ) && val.is_empty()
-        {
-            continue;
-        }
-        let vec_data = _get_json_vec(profile, &val.to_string());
-        if (!key_add_to_details.contains(&key.as_str())
-            && !matches!(
-                profile,
-                Profile::AllFieldInfo(_) | Profile::ExtraFieldInfo(_)
-            ))
-            && vec_data.is_empty()
-        {
-            // A `Details` value of "-" is emitted as an empty object.
-            if matches!(profile, Profile::Details(_)) && val == "-" {
-                record.insert(key.to_string(), JsonNode::Object(IndexMap::new()));
-                continue;
-            }
-            // Plain scalar field. `_convert_valid_json_str` used to drop everything before the
-            // first ": " for multi-segment values (keeping `input[1..]`); reproduce that here.
-            let tmp_val: Vec<&str> = val.split(": ").collect();
-            let joined = if tmp_val.len() == 1 {
-                val.to_string()
-            } else {
-                tmp_val[1..].join(": ")
-            };
-            record.insert(key.to_string(), JsonNode::Leaf(json_scalar(joined.trim())));
-        } else {
-            match profile {
-                // GeoIP profile fields are skipped here because they are emitted inside the
-                // Details/AllFieldInfo sections instead (see the key_add_to_details handling
-                // below).
-                Profile::SrcASN(_)
-                | Profile::SrcCountry(_)
-                | Profile::SrcCity(_)
-                | Profile::TgtASN(_)
-                | Profile::TgtCountry(_)
-                | Profile::TgtCity(_) => continue,
-                Profile::RecoveredRecord(data) => {
-                    record.insert(
-                        "RecoveredRecord".to_string(),
-                        JsonNode::Leaf(json_scalar(data)),
-                    );
-                }
-                Profile::Details(_) | Profile::AllFieldInfo(_) | Profile::ExtraFieldInfo(_) => {
-                    let details_key = match profile {
-                        Profile::Details(_) => "Details",
-                        Profile::AllFieldInfo(_) => "AllFieldInfo",
-                        Profile::ExtraFieldInfo(_) => "ExtraFieldInfo",
-                        _ => "",
-                    };
-                    let details_target_stocks = detect_info
-                        .details_convert_map
-                        .get(&CompactString::from(format!("#{details_key}")));
-                    if details_target_stocks.is_none() {
-                        continue;
-                    }
-                    let details_target_stock = details_target_stocks.unwrap();
-                    // Group the "key: value" detail entries into ordered (key -> values).
-                    let grouped = if detect_info.agg_result.is_some() {
-                        if details_target_stock.is_empty() || details_target_stock[0] == "-" {
-                            record.insert(key.to_string(), JsonNode::Object(IndexMap::new()));
-                            continue;
-                        }
-                        let split_agg_details: Vec<CompactString> = details_target_stock[0]
-                            .split(" ¦ ")
-                            .map(|x| x.into())
-                            .collect();
-                        group_details_raw(&split_agg_details)
-                    } else if details_target_stock.is_empty() {
-                        record.insert(key.to_string(), JsonNode::Object(IndexMap::new()));
-                        continue;
-                    } else {
-                        group_details_raw(details_target_stock)
-                    };
-                    let mut details_obj: IndexMap<String, JsonNode> = IndexMap::new();
-                    for (c_key, c_vals) in grouped {
-                        let node = if c_vals.len() == 1 {
-                            JsonNode::Leaf(json_scalar(&c_vals[0]))
-                        } else {
-                            // A field that repeats becomes an array of string values.
-                            JsonNode::Leaf(serde_json::Value::Array(
-                                c_vals.iter().map(|v| json_string(v)).collect(),
-                            ))
-                        };
-                        details_obj.insert(c_key, node);
-                    }
-                    // Fold GeoIP enrichment fields into Details/AllFieldInfo (not ExtraFieldInfo).
-                    if is_included_geo_ip && !matches!(profile, Profile::ExtraFieldInfo(_)) {
-                        for target_key in valid_key_add_to_details.iter() {
-                            let geo_val = ext_field_map
-                                .get(&CompactString::from(*target_key))
-                                .unwrap()
-                                .to_value();
-                            details_obj.insert(
-                                target_key.to_string(),
-                                JsonNode::Leaf(json_scalar(&geo_val)),
-                            );
-                        }
-                    }
-                    record.insert(key.to_string(), JsonNode::Object(details_obj));
-                }
-                Profile::MitreTags(_) | Profile::MitreTactics(_) | Profile::OtherTags(_) => {
-                    let values: Vec<&str> = val.split(": ").filter(|x| x.trim() != "").collect();
-                    if values.is_empty() {
-                        continue;
-                    }
-                    // Each "value" may itself hold several `¦`-separated tags; flatten them all
-                    // into one array of trimmed strings.
-                    let mut elems: Vec<serde_json::Value> = vec![];
-                    for tag_val in values {
-                        for part in tag_val.split('¦') {
-                            elems.push(json_string(part.trim()));
-                        }
-                    }
-                    record.insert(
-                        key.to_string(),
-                        JsonNode::Leaf(serde_json::Value::Array(elems)),
-                    );
-                }
-                _ => {}
-            }
-        }
-    }
-    (
-        serialize_record_body(record, jsonl_output_flag),
-        next_prev_message,
-    )
-}
-/// Prints a table of the detected rule authors and the number of their rules that produced
-/// detections, laid out over `table_column_num` columns.
-fn output_detected_rule_authors(
-    rule_author_counter: &HashMap<CompactString, i128>,
-    table_column_num: usize,
-) {
-    let mut sorted_authors: Vec<(&CompactString, &i128)> = rule_author_counter.iter().collect();
-
-    // Count descending, then author name ascending for a deterministic order
-    // among equal-count authors.
-    sorted_authors.sort_by(|a, b| b.1.cmp(a.1).then_with(|| a.0.cmp(b.0)));
-    let authors_num = sorted_authors.len();
-    let div = if authors_num <= table_column_num {
-        1
-    } else if !authors_num.is_multiple_of(4) {
-        authors_num / table_column_num + 1
-    } else {
-        authors_num / table_column_num
-    };
-    let mut tb = Table::new();
-    tb.load_preset(UTF8_FULL)
-        .apply_modifier(UTF8_ROUND_CORNERS)
-        .set_style(TableComponent::VerticalLines, ' ');
-    let mut stored_by_column = vec![];
-    let horizontal_line_char = tb.style(TableComponent::HorizontalLines).unwrap();
-    let top_border_char = tb.style(TableComponent::TopBorder).unwrap();
-    for x in 0..table_column_num {
-        let mut tmp = Vec::new();
-        for y in 0..div {
-            if y * table_column_num + x < sorted_authors.len() {
-                // Limit length to 27 to prevent the table from wrapping
-                let filter_author = if sorted_authors[y * table_column_num + x].0.len() <= 27 {
-                    sorted_authors[y * table_column_num + x].0.to_string()
-                } else {
-                    format!("{}...", &sorted_authors[y * table_column_num + x].0[0..24])
-                };
-                tmp.push(format!(
-                    "{} ({})",
-                    filter_author,
-                    sorted_authors[y * table_column_num + x].1
-                ));
-            }
-        }
-        if !tmp.is_empty() {
-            stored_by_column.push(tmp);
-        }
-    }
-    let mut output = vec![];
-    for col_data in stored_by_column {
-        output.push(col_data.join("\n"));
-    }
-    if !output.is_empty() {
-        tb.add_row(output)
-            .set_style(TableComponent::MiddleIntersections, horizontal_line_char)
-            .set_style(TableComponent::TopBorderIntersections, top_border_char)
-            .set_style(
-                TableComponent::BottomBorderIntersections,
-                horizontal_line_char,
-            );
-    }
-    println!("{tb}");
-}
-
-/// Extracts the individual author names from a rule's author string and returns them as an
-/// array.
-fn extract_author_name(author: &str) -> Nested<String> {
-    let mut ret = Nested::<String>::new();
-    for author in author.split(',').map(|s| {
-        // Keep only the part before '(': a parenthesized remark after a name is a description,
-        // not part of the name. Double and single quotes are stripped from the names below.
-        s.split('(').next().unwrap_or_default().to_string()
-    }) {
-        ret.extend(author.split(';'));
-    }
-
-    ret.iter()
-        .map(|r| {
-            r.split('/')
-                .map(|p| p.trim().replace(['"', '\''], ""))
-                .collect::<String>()
-        })
-        .collect()
-}
-
-/// Appends to `html_output_stock` a Markdown table of computer names and the MITRE ATT&CK
-/// tactics detected on them (with unique and total counts), for the HTML report.
-fn _output_html_computer_by_mitre_attck(html_output_stock: &mut Nested<String>) {
-    html_output_stock.push("### MITRE ATT&CK Tactics:{#computers_with_mitre_attck_detections}");
-    if COMPUTER_MITRE_ATTCK_MAP.is_empty() {
-        html_output_stock.push("- No computers were detected with MITRE ATT&CK Tactics.<br>Make sure you run Hayabusa with a profile that includes %MitreTactics% in order to get this info.<br>");
-    }
-    for (idx, sorted_output_map) in COMPUTER_MITRE_ATTCK_MAP
-        .iter()
-        .sorted_by(|a, b| {
-            Ord::cmp(
-                &format!("{}-{}", &b.value()[b.value().len() - 1].0, b.key()),
-                &format!("{}-{}", &a.value()[a.value().len() - 1].0, a.key()),
-            )
-        })
-        .enumerate()
-    {
-        if idx == 0 {
-            html_output_stock.push("|Computer| MITRE ATT&CK Tactics|");
-            html_output_stock.push("|---|---|");
-        }
-        html_output_stock.push(format!(
-            "|{}|{}|",
-            html_escape_value(sorted_output_map.key()),
-            sorted_output_map
-                .value()
-                .iter()
-                .map(|(tactic, unique, total)| format!("{} ({} &#124; {})", tactic, unique, total))
-                .join("<br>")
-        ));
-    }
-    // Scope the accumulated counts to this single report: clear both accumulators after the table is
-    // emitted so a subsequent report generated in the same process (e.g. across tests) starts clean
-    // instead of leaking keys and undercounting `unique`.
-    COMPUTER_MITRE_ATTCK_MAP.clear();
-    COMPUTER_MITRE_ATTCK_UNIQUE_KEYS.clear();
 }
 
 #[cfg(test)]
